Rename GeneralOrdering to Orderer and make it no longer implement Ordering
