+++ src/main/java/org/junit/internal/runners/JUnit38ClassRunner.java
@@ -15,7 +15,7 @@ import org.junit.runner.Description;
 import org.junit.runner.Runner;
 import org.junit.runner.manipulation.Filter;
 import org.junit.runner.manipulation.Filterable;
-import org.junit.runner.manipulation.GeneralOrdering;
+import org.junit.runner.manipulation.Orderer;
 import org.junit.runner.manipulation.InvalidOrderingException;
 import org.junit.runner.manipulation.NoTestsRemainException;
 import org.junit.runner.manipulation.Orderable;
@@ -179,10 +179,10 @@ public class JUnit38ClassRunner extends Runner implements Filterable, Orderable 
      *
      * @since 4.13
      */
-    public void order(GeneralOrdering ordering) throws InvalidOrderingException {
+    public void order(Orderer orderer) throws InvalidOrderingException {
         if (getTest() instanceof Orderable) {
             Orderable adapter = (Orderable) getTest();
-            adapter.order(ordering);
+            adapter.order(orderer);
         }
     }
