+++ src/test/java/org/junit/tests/manipulation/OrderableTest.java
@@ -10,7 +10,7 @@ import org.junit.runner.JUnitCore;
 import org.junit.runner.Request;
 import org.junit.runner.RunWith;
 import org.junit.runner.Runner;
-import org.junit.runner.manipulation.GeneralOrdering;
+import org.junit.runner.manipulation.Orderer;
 import org.junit.runner.manipulation.InvalidOrderingException;
 import org.junit.runner.manipulation.Orderable;
 import org.junit.runner.manipulation.Sorter;
@@ -142,8 +142,8 @@ public class OrderableTest {
                 return delegate.getDescription();
             }
 
-            public void order(GeneralOrdering ordering) throws InvalidOrderingException {
-                delegate.order(ordering);
+            public void order(Orderer orderer) throws InvalidOrderingException {
+                delegate.order(orderer);
             }
 
             public void sort(Sorter sorter) {
